Fix game still running on application exit
The game would continue to run upon closing the game window due to the 'gameLoop' thread still being alive. Added a new boolean to replace the 'gameLoop.IsAlive' as the while loops condition and added a new method 'Exit' to control the exiting of the application. When 'Exit' is called 'isGameActive' is set to false and 'Application.Exit' is called.